refactor(input): fix gamepad open logic, strengthen id wrappers, and improve box geometry API
- Swapped incorrect `SDL_OpenGamepad` / `SDL_GetGamepadFromID` order in `gamepad` constructor.
- Updated `mouse::id` and `sensor::id` constructors to assert validity and use `constexpr` native wrappers.
- Changed `scancode_t` underlying type from `uint16_t` to `uint32_t` to match keyboard code size.
- Rewrote several `box` geometry methods (`composite`, `encloses`, `enclosing`, and line clipping`) for correct optional handling, clipping logic, and type consistency.
- Adjusted joystick/gamepad `all()` implementations to return proper wrapper objects instead of raw IDs.

+++ include/slickdl/scancode.hpp
@@ -8,7 +8,8 @@
 
 namespace slickdl {
 
-using scancode_t = enum class scancode : uint16_t {
+// NOTE: Needs to be ::keyboard::code size
+using scancode_t = enum class scancode : uint32_t {
     unknown = 0,
 
     //  These values are from usage page 0x07 (USB keyboard page)
